Added error handling and validation of Student

+++ src/main/java/com/habbal/demo/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.habbal.demo.student;
 
+import com.habbal.demo.student.exception.BadRequestException;
+import com.habbal.demo.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,10 +18,18 @@ public class StudentService {
     }
 
     public void addStudent(Student student) {
+        if(studentRepository.findByEmail(student.getEmail()).isPresent()) {
+            throw new BadRequestException("Student with email: [" + student.getEmail() + "] already exists");
+        }
+
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
+        if(studentRepository.findById(studentId).isEmpty()) {
+            throw new StudentNotFoundException("Student with ID: [" + studentId + "] does not exist" );
+        }
+
         studentRepository.deleteById(studentId);
     }
 }
